requirements.txt
flask
flask_login
google-cloud-firestore
werkzeug
geopy
pandas
requests